Add exif image validation
* Open and get the exif data during image upload

* Clean the core utils init file

* Change the file validator into cleaner

* Update changelog

* Update validate_slug_and_generate_if_needed tests
