+++ saleor/graphql/product/tests/test_category.py
@@ -1,6 +1,6 @@
 import json
 import os
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, Mock, patch
 
 import graphene
 import pytest
@@ -391,7 +391,8 @@ CATEGORY_CREATE_MUTATION = """
 def test_category_create_mutation(
     monkeypatch, staff_api_client, permission_manage_products, media_root
 ):
-    query = CATEGORY_CREATE_MUTATION
+    # given
+    staff_api_client.user.user_permissions.add(permission_manage_products)
 
     category_name = "Test category"
     description = "description"
@@ -408,10 +409,10 @@ def test_category_create_mutation(
         "backgroundImageAlt": image_alt,
         "slug": category_slug,
     }
-    body = get_multipart_request_body(query, variables, image_file, image_name)
-    response = staff_api_client.post_multipart(
-        body, permissions=[permission_manage_products]
+    body = get_multipart_request_body(
+        CATEGORY_CREATE_MUTATION, variables, image_file, image_name
     )
+    response = staff_api_client.post_multipart(body)
     content = get_graphql_content(response)
     data = content["data"]["categoryCreate"]
     assert data["errors"] == []
@@ -436,7 +437,7 @@ def test_category_create_mutation(
         "parentId": parent_id,
         "slug": f"{category_slug}-2",
     }
-    response = staff_api_client.post_graphql(query, variables)
+    response = staff_api_client.post_graphql(CATEGORY_CREATE_MUTATION, variables)
     content = get_graphql_content(response)
     data = content["data"]["categoryCreate"]
     assert data["errors"] == []
@@ -456,6 +457,8 @@ def test_category_create_trigger_webhook(
     media_root,
     settings,
 ):
+    staff_api_client.user.user_permissions.add(permission_manage_products)
+
     query = CATEGORY_CREATE_MUTATION
     mocked_get_webhooks_for_event.return_value = [any_webhook]
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
@@ -476,9 +479,7 @@ def test_category_create_trigger_webhook(
         "slug": category_slug,
     }
     body = get_multipart_request_body(query, variables, image_file, image_name)
-    response = staff_api_client.post_multipart(
-        body, permissions=[permission_manage_products]
-    )
+    response = staff_api_client.post_multipart(body)
     content = get_graphql_content(response)
     data = content["data"]["categoryCreate"]
     category = Category.objects.first()
@@ -605,6 +606,9 @@ MUTATION_CATEGORY_UPDATE_MUTATION = """
 def test_category_update_mutation(
     monkeypatch, staff_api_client, category, permission_manage_products, media_root
 ):
+    # given
+    staff_api_client.user.user_permissions.add(permission_manage_products)
+
     # create child category and test that the update mutation won't change
     # it's parent
     child_category = category.children.create(name="child")
@@ -629,9 +633,9 @@ def test_category_update_mutation(
     body = get_multipart_request_body(
         MUTATION_CATEGORY_UPDATE_MUTATION, variables, image_file, image_name
     )
-    response = staff_api_client.post_multipart(
-        body, permissions=[permission_manage_products]
-    )
+
+    # when
+    response = staff_api_client.post_multipart(body)
     content = get_graphql_content(response)
     data = content["data"]["categoryUpdate"]
     assert data["errors"] == []
@@ -661,6 +665,8 @@ def test_category_update_trigger_webhook(
     media_root,
     settings,
 ):
+    staff_api_client.user.user_permissions.add(permission_manage_products)
+
     mocked_get_webhooks_for_event.return_value = [any_webhook]
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
 
@@ -683,9 +689,7 @@ def test_category_update_trigger_webhook(
     body = get_multipart_request_body(
         MUTATION_CATEGORY_UPDATE_MUTATION, variables, image_file, image_name
     )
-    response = staff_api_client.post_multipart(
-        body, permissions=[permission_manage_products]
-    )
+    response = staff_api_client.post_multipart(body)
     content = get_graphql_content(response)
     data = content["data"]["categoryUpdate"]
     assert data["errors"] == []
@@ -720,6 +724,8 @@ def test_category_update_background_image_mutation(
     site_settings,
 ):
     # given
+    staff_api_client.user.user_permissions.add(permission_manage_products)
+
     alt_text = "Alt text for an image."
     background_mock = MagicMock(spec=File)
     background_mock.name = "image.jpg"
@@ -754,9 +760,7 @@ def test_category_update_background_image_mutation(
     )
 
     # when
-    response = staff_api_client.post_multipart(
-        body, permissions=[permission_manage_products]
-    )
+    response = staff_api_client.post_multipart(body)
 
     # then
     content = get_graphql_content(response)
@@ -777,7 +781,7 @@ def test_category_update_background_image_mutation(
 
 
 @patch("saleor.core.tasks.delete_from_storage_task.delay")
-def test_category_update_mutation_invalid_background_image(
+def test_category_update_mutation_invalid_background_image_content_type(
     delete_from_storage_task_mock,
     staff_api_client,
     category,
@@ -814,6 +818,58 @@ def test_category_update_mutation_invalid_background_image(
     data = content["data"]["categoryUpdate"]
     assert data["errors"][0]["field"] == "backgroundImage"
     assert data["errors"][0]["message"] == "Invalid file type."
+
+    # ensure that thumbnails for old background image hasn't been deleted
+    assert Thumbnail.objects.filter(category_id=category.id)
+    delete_from_storage_task_mock.assert_not_called()
+
+
+@patch("saleor.core.tasks.delete_from_storage_task.delay")
+def test_category_update_mutation_invalid_background_image(
+    delete_from_storage_task_mock,
+    monkeypatch,
+    staff_api_client,
+    category,
+    permission_manage_products,
+    media_root,
+):
+    # given
+    staff_api_client.user.user_permissions.add(permission_manage_products)
+
+    image_file, image_name = create_image()
+    image_alt = "Alt text for an image."
+
+    error_msg = "Test syntax error"
+    image_file_mock = Mock(side_effect=SyntaxError(error_msg))
+    monkeypatch.setattr(
+        "saleor.graphql.core.validators.file.Image.open", image_file_mock
+    )
+
+    size = 128
+    thumbnail_mock = MagicMock(spec=File)
+    thumbnail_mock.name = "thumbnail_image.jpg"
+    Thumbnail.objects.create(category=category, size=size, image=thumbnail_mock)
+
+    variables = {
+        "name": "new-name",
+        "slug": "new-slug",
+        "id": to_global_id("Category", category.id),
+        "backgroundImage": image_name,
+        "backgroundImageAlt": image_alt,
+        "isPublished": True,
+    }
+    body = get_multipart_request_body(
+        MUTATION_CATEGORY_UPDATE_MUTATION, variables, image_file, image_name
+    )
+
+    # when
+    response = staff_api_client.post_multipart(body)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["categoryUpdate"]
+    assert data["errors"][0]["field"] == "backgroundImage"
+    assert error_msg in data["errors"][0]["message"]
 
     # ensure that thumbnails for old background image hasn't been deleted
     assert Thumbnail.objects.filter(category_id=category.id)
